requirements.txt
pyyaml>=6.0
jsonschema>=4.0
requests>=2.28
